Code for Hackathon
I initially tried to code the roulette wheel but found it too difficult and moved on to do tournament selection. This a failure since I didn't complete the roulette code. I tried to understand it through taking notes, using sudo code, and working with the professor. I tried to take one step at a time but my brain couldn't wrap around the idea. After moving to the tournament selection, it was easier and I was able to finish. My commented out roulette code is still in the file underneath the tournament code.